// src/api/generated/src/models/TransactionFilterDto.ts
export interface TransactionFilterDto {
    /**
     * 
     * @type {number}
     * @memberof TransactionFilterDto
     */
    bankSenderId?: number;
    /**
     * 
     * @type {number}
     * @memberof TransactionFilterDto
     */
    bankReceiverId?: number;
    /**
     * 
     * @type {Date}
     * @memberof TransactionFilterDto
     */
    specificDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof TransactionFilterDto
     */
    dateFrom?: Date;
    /**
     * 
     * @type {Date}
     * @memberof TransactionFilterDto
     */
    dateTo?: Date;
    /**
     * 
     * @type {string}
     * @memberof TransactionFilterDto
     */
    status?: TransactionFilterDtoStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof TransactionFilterDto
     */
    innReceiver?: string;
    /**
     * 
     * @type {number}
     * @memberof TransactionFilterDto
     */
    amountFrom?: number;
    /**
     * 
     * @type {number}
     * @memberof TransactionFilterDto
     */
    amountTo?: number;
    /**
     * 
     * @type {string}
     * @memberof TransactionFilterDto
     */
    transactionType?: TransactionFilterDtoTransactionTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof TransactionFilterDto
     */
    categoryId?: number;
    /**
     * 
     * @type {boolean}
     * @memberof TransactionFilterDto
     */
    anyFilterSet?: boolean;
}


/**
 * @export
 */
export const TransactionFilterDtoStatusEnum = {
    New: 'NEW',
    Confirmed: 'CONFIRMED',
    Processing: 'PROCESSING',
    Canceled: 'CANCELED',
    Completed: 'COMPLETED',
    Deleted: 'DELETED',
    Returned: 'RETURNED'
} as const;
export type TransactionFilterDtoStatusEnum = typeof TransactionFilterDtoStatusEnum[keyof typeof TransactionFilterDtoStatusEnum];

/**
 * @export
 */
export const TransactionFilterDtoTransactionTypeEnum = {
    Income: 'INCOME',
    Expense: 'EXPENSE'
} as const;
export type TransactionFilterDtoTransactionTypeEnum = typeof TransactionFilterDtoTransactionTypeEnum[keyof typeof TransactionFilterDtoTransactionTypeEnum];